Update assignment.py
Made some improvements to the Demolish building function that builds upon the previous code. The current code possesses the ability to check whether the board is empty or whether the chosen area on the board has a building or not before demolishing a building. I chose to kept the previous edition there in case we need to revert to the previous edition.